Fix navigation looping to return unique results for each step.
Previous to this commit, `ORKTaskViewController` only tracked the *last* result, inserting that result into the returned array of results for each instance of the step result with that identifier. This commit fixes that bug so that an `ORKNavigableOrderedTask` with a looping navigation will return the unique result for each loop.